journal: short-circuit stdout_stream_scan()
When s->length is zero this function doesn't do anything.  Simply
short-circuit the function when this is the case.  This is especially
helpful now as it suppresses an unnecessary journal_meta_refresh() when
the length is zero, a common occurrence in force_flush == true (eof)
invocations.

It also clarifies the code, since you must study the implementation to
verify this fact without the explicit short-circuit being present.