Use correct syslog facility number in journald-audit
The hardcoded value of "32" for SYSLOG_FACILITY is not valid. This value
has probably been taken in the bit shifted form, as it is defined in
syslog.h for facility LOG_AUTH.

In case of any integrations (rsyslog imjournal?) the value passed on
would likely be shifter even further, resulting in an incorrect syslog
priority propagated.