Allow ata_id to work for more devices
The current ATA IDENTIFY DEVICE command implementation in ata_id sets CK_COND
bit to 1 in the command and expects the sense data to be returned along with
the response and depends on the sense data for error checking.

Some device including my Jmicron USB bridge (ID 152d:0567 JMicron Technology
Corp.) do not seem to return sense data when CK_COND bit is set to 1 in the
command. Even though it responds successfully to the ATA IDENTIFY DEVICE
command and we have all the data that we need for ata_id to work correctly.

This fix sets CK_COND bit to zero in the command and therefore does not expect
the sense data to be returned unless an error ocurred. Also, we use the regular
SG IO hdr status checks for checking error status instead of checking the sense
data directly. That's the way smartctl seems to handle identify device
commands. This way the command can work sucessfully for more devices.

Fixes #2362