tty-ask-password: Split out password sending
This is a piece of refactoring I've done while looking for a solution to bug #2378.
It separates the password sending from `parse_password`, which only needs to know
about the socket path and the list of passwords to send.

As a caveat, the `ask_password_tty` path needs to construct a one-password strv, too.