dhcp: delay restarts after NAKs
The server might answer to a DHCPREQUEST with a NAK and currently the
client restarts the configuration process immediately.  It was observed
that this can easily generate loops in which the network is flooded with
DISCOVER,OFFER,REQUEST,NAK sequences.

RFC 2131 only states that "if the client receives a DHCPNAK message, the
client restarts the configuration process" without further details, but
also says:

  The client begins in INIT state and forms a DHCPDISCOVER message.
  The client SHOULD wait a random time between one and ten seconds to
  desynchronize the use of DHCP at startup.

It is unclear from the except above if the random delay should be
applied only at startup or every time a DISCOVER is sent; however it
makes sense to delay restarts after the first NAK for a bit so that we
avoid excessive flooding.