journal: introduce mmap_cache_private_to_shared_fd()
mmap_cache_private_to_shared_fd() is used to transition all established
private windows for a given file descriptor to shared windows of the
same file descriptor.  This is used to make journal changes being staged
in memory eligible for journal writeback and visible to the rest of the
system.

The immediate purpose is to facilitate journald continuously executing
for the duration of journal_file_set_offline()'s fsync() calls which
will be moved to a thread.