journal: async journal_file_set_offline() fsync()
Perform fsync() in a separate thread by switching to MAP_PRIVATE
copy-on-write mappings for the duration of the fsync().

This allows the journald event loop to continue executing while a single
offline fsync() occurs in the background.

The fsync() thread is joined and errors handled on the next
journal_file_set_offline() via the new function
journal_file_set_offline_finalize().  journal_file_append_entry() has
also been made a caller of journal_file_set_offline_finalize() in a
non-blocking mode when an fsync is in-progress, to cleanup the private
map and realize any errors before the next offline occurs.

journal_file_close() also finalizes the offline, so the fsync on close
in rotate isn't made asynchronous in these changes.

In a simple test with SyncIntervalSec set to 1 second, the following
latency improvements have been observed:

Now:

  log2(us)   0 1 2 3  4   5  6  7  8 9    10  11  12 13 14 15 16 17
            -------------------------------------------------------
 iterations: 0 2 0 66 156 27 63 9  1 3507 935 48  24 8  2  2  0  1
 iterations: 0 3 1 98 179 28 5  1  1 4059 544 76  28 16 6  1  0  0
 iterations: 1 3 1 34 142 44 55 10 0 3499 687 70  28 9  7  0  1  2
 iterations: 1 1 1 73 164 42 6  2  0 3617 587 81  30 11 2  2  0  1
 iterations: 0 3 2 51 144 44 61 7  0 4029 568 28  27 5  4  0  1  1
 iterations: 0 1 1 74 189 41 13 3  0 4244 478 95  19 11 4  2  0  0
 iterations: 0 2 0 25 125 53 81 11 1 3832 500 76  39 10 7  1  1  1
 iterations: 0 0 0 66 171 39 35 8  0 4153 555 26  13 6  3  0  1  1
 iterations: 0 2 1 28 145 41 83 4  0 3873 590 89  28 12 6  1  1  0
 iterations: 0 2 1 24 140 40 85 13 1 3875 553 103 29 6  5  1  0  1
 iterations: 0 3 0 78 183 27 15 3  0 4029 545 28  27 11 7  0  2  0
 iterations: 0 4 0 77 183 27 12 2  0 3771 578 130 42 12 5  1  0  0

Before:

  log2(us)   0 1 2 3  4   5  6  7  8 9    10  11  12 13 14 15 16 17
            -------------------------------------------------------
 iterations: 0 0 1 78 180 27 17 2  0 4060 711 8   4  0  3  1  1  1
 iterations: 0 0 0 67 170 39 42 2  1 4442 463 9   1  1  1  2  3  0
 iterations: 0 0 0 71 207 26 3  4  0 4189 585 3   4  1  1  0  3  1
 iterations: 0 0 0 72 184 31 21 2  0 4194 533 8   6  0  1  3  2  1
 iterations: 0 0 1 60 186 41 29 4  0 4386 519 3   3  2  0  0  4  0
 iterations: 0 0 0 45 180 28 41 12 0 4050 600 3   9  3  1  1  3  1
 iterations: 0 0 0 77 187 48 5  1  0 4253 632 16  5  0  2  2  2  0
 iterations: 0 0 0 31 145 45 78 12 0 4224 505 3   3  1  1  0  4  1
 iterations: 0 0 0 43 163 27 73 6  0 4275 505 20  6  0  2  1  2  1
 iterations: 0 0 0 72 185 36 28 3  0 4488 452 5   3  0  2  1  4  0
 iterations: 0 1 0 71 201 28 6  1  1 4104 576 15  8  2  0  0  3  1
 iterations: 0 0 0 63 189 29 16 5  1 4083 503 23  10 1  1  1  2  2

In these tests a rotate is occurring in most samples which accounts for the
17th order iterations in both test.

Note however the 16th order has been reduced, instead we see more 11-14th
order iterations.  The current implementation has a tendency to encounter
ENOMEM in mremap() due to not specifying MREMAP_MAYMOVE, which can probably
be embraced by switching to relative pointers.  When we see ENOMEM we
finalize the offline, essentially regressing into the old world of fsync()
stalling the event loop.  This implementation is still an improvement even
in those scenarios however, because we've done some work while the fsync
made progress in the other thread before finalize blocked on waiting for
it to complete.  Such occurrences are represented in those 11-14th order
iterations.