Fix memcpy with NULL pointer arg in base64_append_width
We would call memcpy(x, NULL, 0). This worked fine under -O0,
but did something strange under -O3. This appears illegal:

ISO/IEC 9899:1999 §7.21.1/2 says:
Where an argument declared as size_t n specifies the length of the array
for a function, n can have the value zero on a call to that
function. Unless explicitly stated otherwise in the description of a
particular function in this subclause, pointer arguments on such a call
shall still have valid values, as described in 7.1.4.